Remove GpuDataset related files.
