--- jvm-packages/xgboost4j-spark/src/main/java/ml/dmlc/xgboost4j/java/spark/rapids/PartitionReader.java
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ml.dmlc.xgboost4j.java.spark.rapids;
-
-import java.io.Closeable;
-import java.io.IOException;
-
-/**
- * A partition reader returned by { PartitionReaderFactory#createReader(InputPartition)} or
- * { PartitionReaderFactory#createColumnarReader(InputPartition)}. It's responsible for
- * outputting data for a RDD partition.
- *
- * Note that, Currently the type `T` can only be GpuColumnBatch
- * for normal data sources, or { org.apache.spark.sql.vectorized.ColumnarBatch} for columnar
- * data sources(whose { PartitionReaderFactory#supportColumnarReads(InputPartition)}
- * returns true).
- */
-
-public interface PartitionReader<T> extends Closeable {
-
-  /**
-   * Proceed to next record, returns false if there is no more records.
-   *
-   * @throws IOException if failure happens during disk/network IO like reading files.
-   */
-  boolean next() throws IOException;
-
-  /**
-   * Return the current record. This method should return same value until `next` is called.
-   */
-  T get();
-}
