+++ jvm-packages/xgboost4j-spark/src/main/scala/ml/dmlc/xgboost4j/scala/spark/XGBoostClassifier.scala
@@ -20,7 +20,7 @@ import ai.rapids.cudf.{Rmm, RmmAllocationMode, Table}
 import ml.dmlc.xgboost4j.java.spark.rapids.GpuColumnBatch
 import ml.dmlc.xgboost4j.java.{Rabit, XGBoostSparkJNI}
 import ml.dmlc.xgboost4j.scala.spark.params._
-import ml.dmlc.xgboost4j.scala.spark.rapids.{ColumnBatchToRow, GpuDataset, PluginUtils}
+import ml.dmlc.xgboost4j.scala.spark.rapids.PluginUtils
 import ml.dmlc.xgboost4j.scala.{Booster, DMatrix, EvalTrait, ObjectiveTrait, XGBoost => SXGBoost}
 import ml.dmlc.xgboost4j.{LabeledPoint => XGBLabeledPoint}
 import org.apache.commons.logging.LogFactory
@@ -370,7 +370,7 @@ class XGBoostClassificationModel private[ml](
       if (gpuId == 0) gpuId = -1
 
       val columnBatchIter = iter.map(new GpuColumnBatch(_, originalSchema))
-      val ((dm, columnBatchToRow), time) = GpuDataset.time("Transform: build dmatrix and row") {
+      val ((dm, columnBatchToRow), time) = PluginUtils.time("Transform: build dmatrix and row") {
         DataUtils.buildDMatrixIncrementally(gpuId, missing, featureIndices, columnBatchIter)
       }
       logger.debug("Benchmark [Transform: Build Dmatrix and Row] " + time)
