--- jvm-packages/xgboost4j-spark/src/main/scala/ml/dmlc/xgboost4j/scala/spark/rapids/FilePartitionReaderFactory.scala
@@ -1,48 +1,0 @@
-/*
- Copyright (c) 2014 by Contributors
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
- http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
- */
-
-
-package ml.dmlc.xgboost4j.scala.spark.rapids
-
-import ml.dmlc.xgboost4j.java.spark.rapids.{GpuColumnBatch, PartitionReader, PartitionReaderFactory}
-import org.apache.spark.sql.execution.datasources.{FilePartition, PartitionedFile}
-
-abstract class FilePartitionReaderFactory extends PartitionReaderFactory {
-
-  override def createColumnarReader(partition: FilePartition): PartitionReader[GpuColumnBatch] = {
-    assert(partition.isInstanceOf[FilePartition])
-    val filePartition = partition.asInstanceOf[FilePartition]
-    val iter = GpuDataset.getFiles(filePartition).toIterator.map { file =>
-      PartitionedFileReader(file, buildColumnarReader(file))
-    }
-    new FilePartitionReader[GpuColumnBatch](iter)
-  }
-
-  def buildColumnarReader(partitionedFile: PartitionedFile): PartitionReader[GpuColumnBatch] = {
-    throw new UnsupportedOperationException("Cannot create columnar reader.")
-  }
-}
-
-private[this] case class PartitionedFileReader[T](file: PartitionedFile,
-    reader: PartitionReader[T]) extends PartitionReader[T] {
-  override def next(): Boolean = reader.next()
-
-  override def get(): T = reader.get()
-
-  override def close(): Unit = reader.close()
-
-  override def toString: String = file.toString
-}
