+++ jvm-packages/xgboost4j-spark/src/main/scala/ml/dmlc/xgboost4j/scala/spark/rapids/PluginUtils.scala
@@ -36,4 +36,12 @@ object PluginUtils extends Serializable {
       .asInstanceOf[RDD[Table]]
   }
   // scalastyle:on classforname
+
+  // calculate bench mark
+  def time[R](phase: String)(block: => R): (R, Float) = {
+    val t0 = System.currentTimeMillis
+    val result = block // call-by-name
+    val t1 = System.currentTimeMillis
+    (result, (t1 - t0).toFloat / 1000)
+  }
 }
